requirements.txt
Django==4.1.5
psycopg2-binary=2.9.5
django-recaptcha=3.0.0
